youtube_api/requirements.txt
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==2.2.3
djangorestframework==3.10.1
gunicorn==19.9.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
nltk==3.4.4
numpy==1.16.4
pandas==0.25.0
Pillow==6.1.0
psycopg2==2.8.3
pyparsing==2.4.1.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
sqlparse==0.3.0
textblob==0.15.3
Unidecode==1.1.1
urllib3==1.25.3
whitenoise==4.1.3
wordcloud==1.5.0
